requirements.txt
fastapi
uvicorn
flet
sqlalchemy
psycopg2-binary
python-dotenv
requests
pytest
flet-fastapi
